+++ pdfminer/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "20220317"
+__version__ = "20220318"
 
 if __name__ == "__main__":
     print(__version__)