+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes in pdfminer.six will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [20220318]
+
+### Fixed
+
+- Fixed Bugs: Fix font with CIDFont name and cmap empty, 
+  - Fixed cmap without `endbfchar` 
+
 ## [20220317]
 
 ### Fixed